# fix: improve digest64 for as-sha256