--- a/.github/workflows/benchmark.yml
+++ b/.github/workflows/benchmark.yml
@@ -26,7 +26,7 @@ jobs:
         with:
           cache: yarn
           node-version: 18
-      - run: yarn --immutable && yarn build
+      - run: yarn --immutable && yarn build && yarn generate
 
       - name: Run benchmarks
         run: yarn benchmark