--- a/packages/as-sha256/package.json
+++ b/packages/as-sha256/package.json
@@ -34,15 +34,13 @@
     "asbuild:optimized": "asc assembly/index.ts -b build/optimized.wasm -t build/optimized.wat --runtime minimal --target release --noAssert",
     "build:lib": "run -T tsc -p tsconfig.build.json",
     "build:web": "run -T webpack --mode production --entry ./index.js --output ./dist/as-sha256.min.js",
-    "test": "yarn run test:as-ci && yarn run test:unit",
+    "test": "yarn run test:unit",
     "test:unit": "yarn run test:unit:node && yarn run test:unit:browser",
     "test:unit:node": "run -T mocha -r ts-node/register test/unit/*.test.ts",
     "test:unit:browser": "run -T karma start --single-run --browsers ChromeHeadless,FirefoxHeadless karma.config.js",
     "benchmark": "node -r ts-node/register ./node_modules/.bin/benchmark 'test/perf/index.test.ts'",
     "benchmark:local": "yarn benchmark --local",
-    "test:ci": "yarn test:as-ci",
-    "test:as": "asp --nortrace --verbose",
-    "test:as-ci": "asp --nortrace 2> /dev/null"
+    "test:ci": "yarn test:as-ci"
   },
   "devDependencies": {
     "@chainsafe/babel-plugin-inline-binary-import": "^1.0.3",