--- a/packages/as-sha256/package.json
+++ b/packages/as-sha256/package.json
@@ -28,7 +28,7 @@
     "build"
   ],
   "scripts": {
-    "clean": "rm -rf ./dist && node -r ts-node/register ./scripts/codegen.ts",
+    "generate": "rm -rf ./dist && node -r ts-node/register ./scripts/codegen.ts",
     "build": "yarn asbuild:untouched && yarn asbuild:optimized && yarn build:lib",
     "asbuild:untouched": "asc assembly/index.ts -o build/untouched.wasm -t build/untouched.wat --runtime minimal --target debug",
     "asbuild:optimized": "asc assembly/index.ts -o build/optimized.wasm -t build/optimized.wat --runtime minimal --target release -O3z --noAssert",