**Motivation**

Improve `digest64`

**Description**

- Use latest `assemblyscript` yields better performance
- Chain `w` computation to the main loop in `hashBlocks()`
- Return result by using `Uint8Array.slice()`

Before:
```
digestTwoHashObjects vs digest64 vs digest
    ✓ digestTwoHashObjects 50023 times                                    27.71683 ops/s    36.07916 ms/op   x1.042        264 runs   10.1 s
    ✓ digest64 50023 times                                                26.85761 ops/s    37.23339 ms/op   x1.020        256 runs   10.1 s
    ✓ digest 50023 times                                                  27.02430 ops/s    37.00373 ms/op   x1.010        259 runs   10.1 s

```

After
```
digestTwoHashObjects vs digest64 vs digest
    ✓ digestTwoHashObjects 50023 times                                    30.54306 ops/s    32.74066 ms/op   x0.946        293 runs   10.1 s
    ✓ digest64 50023 times                                                28.28273 ops/s    35.35727 ms/op   x0.968        271 runs   10.1 s
    ✓ digest 50023 times                                                  27.93469 ops/s    35.79778 ms/op   x0.977        268 runs   10.1 s
  
```

`digestTwoHashObjects` is ~10% faster and `digest64` is ~5% faster