@tuyennhv 2024-02-28
the benchmark result in CI is not precise because it run with very short period of time, updated it to run at least 10s

```typescript
setBenchOpts({
    minMs: 10000,
  });
```