@tuyennhv 2024-02-28
> Do we need to store `build` artifacts in the repo?

yes, they are actually input for our typescript function every time we run, that's why they are tracked in git