@jeluard 2024-02-28
> > Do we need to store `build` artifacts in the repo?
> 
> yes, they are actually input for our typescript function every time we run, that's why they are tracked in git

Is [this](https://github.com/ChainSafe/ssz/blob/master/packages/as-sha256/scripts/codegen.ts#L7) what you are referring to? Looks like we only care about `codegen` being called to update https://github.com/ChainSafe/ssz/blob/master/packages/as-sha256/src/wasmCode.ts (as correctly done by this PR), but wasm/wat don't need to be in git?

Not part of this PR anyway, maybe something to be considered to improve the build process.